requirements.txt
certifi==2017.4.17
chardet==3.0.4
github3.py==0.9.3
idna==2.5
requests==2.18.1
ruamel.yaml==0.15.19
uritemplate==3.0.0
uritemplate.py==3.0.2
urllib3==1.21.1
